==> 03_code/02_fine_tuning_health.ipynb <==
## Fine-Tuning on Training Dataset for topic "Health"

Meant to be run on a Google Colab Instance, not locally. T4, A100 or V100 should all be sufficient. 

Thanks to Moritz Laurer for excellent templates and starting points for fine-tuning! See [here](https://github.com/MoritzLaurer/summer-school-transformers-2023/blob/main/3_tune_bert.ipynb)

#### Install relevant packages

!pip install accelerate -U
!pip install transformers[sentencepiece]
!pip install datasets


## Load general packages
# some more specialised packages are loaded in each sub section
import pandas as pd
import numpy as np
from google.colab.data_table import DataTable
from sklearn.model_selection import train_test_split

# set random seed for reproducibility
SEED_GLOBAL = 42
np.random.seed(SEED_GLOBAL)

## Download data

df =  pd.read_excel("https://nextcloud.swp-berlin.org/s/redacted/download").dropna(subset = ['text'])

df["label_text"] = df["label"].apply(lambda x: 'health' if x == 1 else 'not health')



df

df = df.drop_duplicates(subset='text', keep='first')
df



df_train, df_test = train_test_split(df, test_size=0.2, random_state=42)

df_train = df_train[["label", "label_text", "text","id"]]

df_test = df_test[["label", "label_text", "text","id"]]
df_test.head()

df_test['text_exists_in_train'] = df_test['text'].isin(df_train['text'])
df_test["text_exists_in_train"].sum()

from transformers import AutoTokenizer, AutoModelForSequenceClassification, AutoConfig
import torch

## load a model and its tokenizer
model_name = "microsoft/deberta-v3-base"  # replace e.g. with "microsoft/deberta-v3-base"
tokenizer = AutoTokenizer.from_pretrained(model_name, use_fast=True, model_max_length=512)


# link the numeric labels to the label texts
label_text = np.sort(df_test.label_text.unique()).tolist()
# Get unique label_text and corresponding label values
unique_labels = df[['label', 'label_text']].drop_duplicates()

label2id = dict(zip(unique_labels['label_text'], unique_labels['label']))
id2label = dict(zip(unique_labels['label'], unique_labels['label_text']))


config = AutoConfig.from_pretrained(model_name, label2id=label2id, id2label=id2label, num_labels=len(label2id))
config.hidden_dropout_prob = 0.02
# load model with config
model = AutoModelForSequenceClassification.from_pretrained(model_name, config=config, ignore_mismatched_sizes=True)

# use GPU (cuda) if available, otherwise use CPU
device = "cuda" if torch.cuda.is_available() else "cpu"
print(f"Device: {device}")
model.to(device);

## Setting training arguments / hyperparameters

### Function to calculate metrics
from sklearn.metrics import balanced_accuracy_score, precision_recall_fscore_support, accuracy_score, classification_report
from transformers import TrainingArguments, Trainer, logging

import warnings

def compute_metrics_standard(eval_pred):
    with warnings.catch_warnings():
        warnings.filterwarnings("ignore")

        labels = eval_pred.label_ids
        pred_logits = eval_pred.predictions
        preds_max = np.argmax(pred_logits, axis=1)  # argmax on each row (axis=1) in the tensor

        # metrics
        precision_macro, recall_macro, f1_macro, _ = precision_recall_fscore_support(labels, preds_max, average='macro')  # https://scikit-learn.org/stable/modules/generated/sklearn.metrics.precision_recall_fscore_support.html
        acc_balanced = balanced_accuracy_score(labels, preds_max)

        metrics = {
            'f1_macro': f1_macro,
            'accuracy_balanced': acc_balanced,
            'precision_macro': precision_macro,
            'recall_macro': recall_macro
        }

        return metrics


from google.colab import drive
drive.mount('/content/drive')

import datasets
dataset = datasets.DatasetDict({
    "train": datasets.Dataset.from_pandas(df_train),
    "test": datasets.Dataset.from_pandas(df_test)
})
print(df_train['text'].iloc[0])

# tokenize
def tokenize(examples):
  return tokenizer(examples["text"], truncation=True, max_length=512)  # max_length can be reduced to e.g. 256 to increase speed, but long texts will be cut off

dataset["train"] = dataset["train"].map(tokenize, batched=True)
dataset["test"] = dataset["test"].map(tokenize, batched=True)
dataset = dataset.remove_columns(['label_text'])


## Fine-tuning and evaluation

Let's start fine-tuning the model!

If you get an 'out-of-memory' error, reduce the 'per_device_train_batch_size' to 8 or 4 in the TrainingArguments above and restart the runtime. If you don't restart your runtime (menu to the to left 'Runtime' > 'Restart runtime') and rerun the entire script, the 'out-of-memory' error will probably not go away.

import numpy as np
import json
import os

## push to hugging face with swp account
## to do learning rate scheduler to constant instead of linear and make sure to do at least 30 epochs
## https://medium.com/geekculture/how-does-batch-size-impact-your-model-learning-2dd34d9fb1fa
## see https://arxiv.org/pdf/2111.09543.pdf

if model_name == "microsoft/deberta-v3-large":
  gradient_accumulation_steps = 16
  LEARNING_RATE = 8e-6  # can try: 9e-6
  EPOCHS = 30  # can try: 10
  BATCH_SIZE = 4  # can try: 10
  folder = "deberta_large"
else:
  LEARNING_RATE = 6e-5  # can try: 6e-5
  EPOCHS = 30  # can try: 10
  BATCH_SIZE = 8  # can try: 10
  folder = "deberta_base"
  gradient_accumulation_steps = 4




train_args = TrainingArguments(
    output_dir = f"./output2/",
    logging_dir=f'./logs/logs',
    gradient_accumulation_steps=gradient_accumulation_steps,
    num_train_epochs=EPOCHS,  # this can be increased, but higher values increase training time. Good values for NLI are between 3 and 20.
    learning_rate=LEARNING_RATE, # with deberta base was 6e-5
    per_device_train_batch_size=BATCH_SIZE,  # if you get an out-of-memory error, reduce this value to 8 or 4 and restart the runtime. Higher values increase training speed, but also increase memory requirements. Ideal values here are always a multiple of 8.
    per_device_eval_batch_size=BATCH_SIZE,  # if you get an out-of-memory error, reduce this value, e.g. to 40 and restart the runtime
    warmup_ratio=0.06,  # a good normal default value is 0.06 for normal BERT-base models, but since we want to reuse prior NLI knowledge and avoid catastrophic forgetting, we set the value higher
    weight_decay=0.01,
    load_best_model_at_end=True,
    metric_for_best_model="f1_macro",
    evaluation_strategy="epoch", # options: "no"/"steps"/"epoch"
    save_strategy = "epoch",  # options: "no"/"steps"/"epoch",
    save_total_limit = 2,
    report_to="all",  # "all"  # logging
    )

# remove unnecessary columns for model training
trainer = Trainer(
    model=model,
    tokenizer=tokenizer,
    args=train_args,
    train_dataset=dataset["train"],
    eval_dataset=dataset["test"],
    compute_metrics=compute_metrics_standard
)

trainer.train(resume_from_checkpoint = False)
eval_metrics = trainer.evaluate()

trainer.evaluate()

run_below = False
assert run_below

### save best model to google drive
model_name_custom = f"deberta-base-health_final_20240321_9596"
mode_custom_path = "/content/drive/MyDrive/unga_health/" + model_name_custom

trainer.save_model(output_dir=mode_custom_path)

df_test.to_csv("/content/drive/MyDrive/unga_health/"+model_name_custom+'_test_data.csv')
