==> 03_code/03_inference_health.ipynb <==
## Inference on Dataset for topic "Health"

Meant to be run on a Google Colab Instance, not locally. T4, A100 or V100 should all be sufficient. 

#### Install relevant packages

!pip install accelerate -U
!pip install transformers[sentencepiece]
!pip install datasets
!pip install --force-reinstall -v "openpyxl==3.0.10"
!pip install xformers

## Load general packages
import pandas as pd
import numpy as np
from google.colab.data_table import DataTable
from sklearn.model_selection import train_test_split
from google.colab import drive
from transformers import pipeline, AutoModelForSequenceClassification, AutoTokenizer
import torch
import os
import tqdm

### Connect to drive

drive.mount('/content/drive')

### Download Model from Drive



model_name_custom = f"deberta-base-health_final_20240321_9596"
mode_custom_path = "/content/drive/MyDrive/unga_health/" + model_name_custom
device = "cuda:0" if torch.cuda.is_available() else "cpu"  # use GPU (cuda) if available, otherwise use CPU

model = AutoModelForSequenceClassification.from_pretrained(mode_custom_path)
tokenizer = AutoTokenizer.from_pretrained(mode_custom_path, use_fast=True, model_max_length=512)

# documentation: https://huggingface.co/docs/transformers/main_classes/pipelines#transformers.ZeroShotClassificationPipeline
pipe_classifier = pipeline(
    "text-classification",
    model=model,  # if you have trained a model above, load_best_model_at_end in the training arguments has automatically replaced model with the fine-tuned model
    tokenizer=tokenizer,
    framework="pt",
    device=device,
    batch_size=24
)

with torch.no_grad():
  torch.cuda.empty_cache()

### Download Data

oos_test = pd.read_csv('https://nextcloud.swp-berlin.org/s/REDACTED/download')
oos_test =  oos_test[['text', 'detail_vote_number', "group_id_alt"]]
oos_test


oos_test['text_length'] = oos_test['text'].str.len()
# Get the top 100 texts by length
top_100_texts_by_length = oos_test.nlargest(100, 'text_length')
top_100_texts_by_length

## drop missing texts, this is because we have all resolutions in the dataset,
## even the ones where the PDF is faulty or we dont have a link
oos_test = oos_test.dropna(subset=['text'])

## delete the text from paragraphs that are simply too long for the classifier
## use only the first 6000 characters of the
oos_test["text"] = oos_test.apply(lambda row: row['text'][:6000] if len(row['text']) > 25000 else row['text'], axis=1)

oos_test

dir_path = '/content/drive/MyDrive/unga_health/preds_20240416'
if not os.path.exists(dir_path):
    os.makedirs(dir_path)

text_lst = oos_test["text"].tolist()

# in principle it would be better to use a Dataset object from the transformers library
# however, that does not allow for saving intermediate results and since it is a lot of text
# sometimes the google colab crashes or gets disconnected. This way it probably takes
# about half an hour longer but that doesn´t matter
def save_output(output, chunk_idx):
    """Save intermediate results to a CSV file."""
    file_path = f"{dir_path}/predictions_eval_set_chunk_{chunk_idx}.csv"
    df_temp = pd.DataFrame([data[0] for data in output])
    eval_temp = pd.concat([oos_test.iloc[chunk_idx*1000:(chunk_idx+1)*1000].reset_index(drop=True), df_temp], axis=1, ignore_index=True)
    eval_temp.to_csv(file_path, index=False)

def process_texts(texts):
    """Process a list of texts in chunks and save outputs."""
    num_chunks = len(texts) // 1000 + (1 if len(texts) % 1000 != 0 else 0)
    with torch.no_grad():
      torch.cuda.empty_cache()

    with tqdm.tqdm(total=len(texts)) as pbar:
        for chunk_idx in range(num_chunks):
            file_path = f"{dir_path}/predictions_eval_set_chunk_{chunk_idx}.csv"

            # Check if the file is already saved
            if os.path.exists(file_path):
                pbar.update(1000)  # Update the progress bar by 1000 steps
                continue

            start_idx = chunk_idx * 1000
            end_idx = start_idx + 1000
            current_chunk_output = [pipe_classifier(text) for text in texts[start_idx:end_idx]]

            save_output(current_chunk_output, chunk_idx)
            pbar.update(1000)  # Update the progress bar by 1000 steps

process_texts(text_lst)

import random

# Set working directory
os.chdir('/content/drive/MyDrive/unga_health/preds_20240321')

# List all csv files in the directory
files = [f for f in os.listdir() if f.endswith('.csv')]

# Read and process all csv files
predictions = pd.concat([pd.read_csv(file) for file in files], ignore_index=True)

predictions.columns = ['text', 'detail_vote_number', 'group_id_alt','index_alt', 'label', 'prob']
predictions = predictions.drop(columns=[predictions.columns[3]])

# Get rows with the lowest 200 probabilities
low_prob = predictions.nsmallest(200, 'prob')



# Filter rows where text contains "health" and label is "not health"
health = predictions[predictions['text'].str.contains('health', case=False) & (predictions['label'] == 'not health')]

# Get rows with the highest 200 probabilities
high_prob = predictions.nlargest(200, 'prob')

# Sample data
coders = ["daniel", "rebecca", "paul"]
sample_list = []

for coder in coders:
    for label in predictions['label'].unique():
        sample_data = predictions[predictions['label'] == label].sample(n=25)
        sample_data['coder'] = coder
        sample_list.append(sample_data)

sample = pd.concat(sample_list, ignore_index=True)

sample


low_prob

health
